refactor: group classes by dimensionality
