+++ src/vector/_d3/__init__.py
@@ -1,0 +1,11 @@
+"""3-dimensional."""
+
+from . import base, builtin, transform
+from .base import *
+from .builtin import *
+from .transform import *
+
+__all__: list[str] = []
+__all__ += base.__all__
+__all__ += builtin.__all__
+__all__ += transform.__all__
