+++ src/vector/_transform.py
@@ -10,48 +10,15 @@ import astropy.units as u
 from jax_quantity import Quantity
 from plum import dispatch
 
-from ._base import (  # pylint: disable=cyclic-import
-    Abstract1DVector,
-    Abstract2DVector,
-    Abstract3DVector,
-)
-from ._builtin import (  # pylint: disable=cyclic-import
-    Cartesian1DVector,
-    Cartesian2DVector,
-    Cartesian3DVector,
-    CylindricalVector,
-    # LnPolarVector,
-    # Log10PolarVector,
-    PolarVector,
-    RadialVector,
-    SphericalVector,
-)
+from ._d1.builtin import Cartesian1DVector, RadialVector
+from ._d2.base import Abstract2DVector  # pylint: disable=cyclic-import
+from ._d2.builtin import Cartesian2DVector, PolarVector
+from ._d3.base import Abstract3DVector
+from ._d3.builtin import Cartesian3DVector, CylindricalVector, SphericalVector
 from ._exceptions import IrreversibleDimensionChange
 
 ###############################################################################
 # 1D
-
-
-@dispatch
-def represent_as(
-    current: Abstract1DVector, target: type[Abstract1DVector], /, **kwargs: Any
-) -> Abstract1DVector:
-    """Abstract1DVector -> Cartesian1D -> Abstract1DVector.
-
-    This is the base case for the transformation of 1D vectors.
-    """
-    cart1d = represent_as(current, Cartesian1DVector)
-    return represent_as(cart1d, target)
-
-
-@dispatch.multi(
-    (Cartesian1DVector, type[Cartesian1DVector]), (RadialVector, type[RadialVector])
-)
-def represent_as(
-    current: Abstract1DVector, target: type[Abstract1DVector], /, **kwargs: Any
-) -> Abstract1DVector:
-    """Self transform of 1D vectors."""
-    return current
 
 
 # @dispatch.multi(
@@ -72,21 +39,6 @@ def represent_as(
 
 # =============================================================================
 # Cartesian1DVector
-
-
-# -----------------------------------------------
-# 1D
-
-
-@dispatch
-def represent_as(
-    current: Cartesian1DVector, target: type[RadialVector], /, **kwargs: Any
-) -> RadialVector:
-    """Cartesian1DVector -> RadialVector.
-
-    The `x` coordinate is converted to the radial coordinate `r`.
-    """
-    return target(r=current.x)
 
 
 # -----------------------------------------------
@@ -223,21 +175,6 @@ def represent_as(
 # RadialVector
 
 # -----------------------------------------------
-# 1D
-
-
-@dispatch
-def represent_as(
-    current: RadialVector, target: type[Cartesian1DVector], /, **kwargs: Any
-) -> Cartesian1DVector:
-    """RadialVector -> Cartesian1DVector.
-
-    The `r` coordinate is converted to the `x` coordinate of the 1D system.
-    """
-    return target(x=current.r)
-
-
-# -----------------------------------------------
 # 2D
 
 
@@ -337,44 +274,6 @@ def represent_as(
 # 2D
 
 
-@dispatch
-def represent_as(
-    current: Abstract2DVector, target: type[Abstract2DVector], /, **kwargs: Any
-) -> Abstract2DVector:
-    """Abstract2DVector -> Cartesian2D -> Abstract2DVector.
-
-    This is the base case for the transformation of 2D vectors.
-    """
-    return represent_as(represent_as(current, Cartesian2DVector), target)
-
-
-@dispatch.multi(
-    (Cartesian2DVector, type[Cartesian2DVector]),
-    (PolarVector, type[PolarVector]),
-    # (LnPolarVector, type[LnPolarVector]),
-    # (Log10PolarVector, type[Log10PolarVector]),
-)
-def represent_as(
-    current: Abstract2DVector, target: type[Abstract2DVector], /, **kwargs: Any
-) -> Abstract2DVector:
-    """Self transform of 2D vectors."""
-    return current
-
-
-# @dispatch.multi(
-#     (Cartesian2DVector, type[LnPolarVector]),
-#     (Cartesian2DVector, type[Log10PolarVector]),
-#     (LnPolarVector, type[Cartesian2DVector]),
-#     (Log10PolarVector, type[Cartesian2DVector]),
-# )
-# def represent_as(
-#     current: Abstract2DVector, target: type[Abstract2DVector], /, **kwargs: Any
-# ) -> Abstract2DVector:
-#     """Abstract2DVector -> PolarVector -> Abstract2DVector."""
-#     polar = represent_as(current, PolarVector)
-#     return represent_as(polar, target)
-
-
 @dispatch.multi(
     (Cartesian2DVector, type[SphericalVector]),
     (Cartesian2DVector, type[CylindricalVector]),
@@ -456,24 +355,6 @@ def represent_as(
 
 
 # -----------------------------------------------
-# 2D
-
-
-@dispatch
-def represent_as(
-    current: Cartesian2DVector, target: type[PolarVector], /, **kwargs: Any
-) -> PolarVector:
-    """Cartesian2DVector -> PolarVector.
-
-    The `x` and `y` coordinates are converted to the radial coordinate `r` and
-    the angular coordinate `phi`.
-    """
-    r = xp.sqrt(current.x**2 + current.y**2)
-    phi = xp.atan2(current.y, current.x)
-    return target(r=r, phi=phi)
-
-
-# -----------------------------------------------
 # 3D
 
 
@@ -520,36 +401,6 @@ def represent_as(
 
 
 # -----------------------------------------------
-# 2D
-
-
-@dispatch
-def represent_as(
-    current: PolarVector, target: type[Cartesian2DVector], /, **kwargs: Any
-) -> Cartesian2DVector:
-    """PolarVector -> Cartesian2DVector."""
-    x = current.r * xp.cos(current.phi)
-    y = current.r * xp.sin(current.phi)
-    return target(x=x, y=y)
-
-
-# @dispatch
-# def represent_as(
-#     current: PolarVector, target: type[LnPolarVector], /, **kwargs: Any
-# ) -> LnPolarVector:
-#     """PolarVector -> LnPolarVector."""
-#     return target(lnr=xp.log(current.r), phi=current.phi)
-
-
-# @dispatch
-# def represent_as(
-#     current: PolarVector, target: type[Log10PolarVector], /, **kwargs: Any
-# ) -> Log10PolarVector:
-#     """PolarVector -> Log10PolarVector."""
-#     return target(log10r=xp.log10(current.r), phi=current.phi)
-
-
-# -----------------------------------------------
 # 3D
 
 
@@ -591,26 +442,6 @@ def represent_as(
 #     return target(r=xp.exp(current.lnr))
 
 
-# # -----------------------------------------------
-# # 2D
-
-
-# @dispatch
-# def represent_as(
-#     current: LnPolarVector, target: type[PolarVector], /, **kwargs: Any
-# ) -> PolarVector:
-#     """LnPolarVector -> PolarVector."""
-#     return target(r=xp.exp(current.lnr), phi=current.phi)
-
-
-# @dispatch
-# def represent_as(
-#     current: LnPolarVector, target: type[Log10PolarVector], /, **kwargs: Any
-# ) -> Log10PolarVector:
-#     """LnPolarVector -> Log10PolarVector."""
-#     return target(log10r=current.lnr / xp.log(10), phi=current.phi)
-
-
 # # =============================================================================
 # # Log10PolarVector
 
@@ -638,48 +469,8 @@ def represent_as(
 #     return target(r=xp.pow(10, current.log10r))
 
 
-# # -----------------------------------------------
-# # 2D
-
-
-# @dispatch
-# def represent_as(
-#     current: Log10PolarVector, target: type[PolarVector], /, **kwargs: Any
-# ) -> PolarVector:
-#     """Log10PolarVector -> PolarVector."""
-#     return target(r=xp.pow(10, current.log10r), phi=current.phi)
-
-
-# @dispatch
-# def represent_as(
-#     current: Log10PolarVector, target: type[LnPolarVector], /, **kwargs: Any
-# ) -> LnPolarVector:
-#     """Log10PolarVector -> LnPolarVector."""
-#     return target(lnr=current.log10r * xp.log(10), phi=current.phi)
-
-
 ###############################################################################
 # 3D
-
-
-@dispatch
-def represent_as(
-    current: Abstract3DVector, target: type[Abstract3DVector], /, **kwargs: Any
-) -> Abstract3DVector:
-    """Abstract3DVector -> Cartesian3D -> Abstract3DVector."""
-    return represent_as(represent_as(current, Cartesian3DVector), target)
-
-
-@dispatch.multi(
-    (Cartesian3DVector, type[Cartesian3DVector]),
-    (SphericalVector, type[SphericalVector]),
-    (CylindricalVector, type[CylindricalVector]),
-)
-def represent_as(
-    current: Abstract3DVector, target: type[Abstract3DVector], /, **kwargs: Any
-) -> Abstract3DVector:
-    """Self transform."""
-    return current
 
 
 # @dispatch.multi(
@@ -751,31 +542,6 @@ def represent_as(
     return represent_as(cart2, target)
 
 
-# -----------------------------------------------
-# 3D
-
-
-@dispatch
-def represent_as(
-    current: Cartesian3DVector, target: type[SphericalVector], /, **kwargs: Any
-) -> SphericalVector:
-    """Cartesian3DVector -> SphericalVector."""
-    r = xp.sqrt(current.x**2 + current.y**2 + current.z**2)
-    theta = xp.acos(current.z / r)
-    phi = xp.atan2(current.y, current.x)
-    return target(r=r, theta=theta, phi=phi)
-
-
-@dispatch
-def represent_as(
-    current: Cartesian3DVector, target: type[CylindricalVector], /, **kwargs: Any
-) -> CylindricalVector:
-    """Cartesian3DVector -> CylindricalVector."""
-    rho = xp.sqrt(current.x**2 + current.y**2)
-    phi = xp.atan2(current.y, current.x)
-    return target(rho=rho, phi=phi, z=current.z)
-
-
 # =============================================================================
 # SphericalVector
 
@@ -826,32 +592,6 @@ def represent_as(
     return target(r=current.r * xp.sin(current.theta), phi=current.phi)
 
 
-# -----------------------------------------------
-# 3D
-
-
-@dispatch
-def represent_as(
-    current: SphericalVector, target: type[Cartesian3DVector], /, **kwargs: Any
-) -> Cartesian3DVector:
-    """SphericalVector -> Cartesian3DVector."""
-    x = current.r * xp.sin(current.theta) * xp.cos(current.phi)
-    y = current.r * xp.sin(current.theta) * xp.sin(current.phi)
-    z = current.r * xp.cos(current.theta)
-    return target(x=x, y=y, z=z)
-
-
-@dispatch
-def represent_as(
-    current: SphericalVector, target: type[CylindricalVector], /, **kwargs: Any
-) -> CylindricalVector:
-    """SphericalVector -> CylindricalVector."""
-    rho = current.r * xp.sin(current.theta)
-    phi = current.phi
-    z = current.r * xp.cos(current.theta)
-    return target(rho=rho, phi=phi, z=z)
-
-
 # =============================================================================
 # CylindricalVector
 
@@ -900,29 +640,3 @@ def represent_as(
     """CylindricalVector -> PolarVector."""
     warn("irreversible dimension change", IrreversibleDimensionChange, stacklevel=2)
     return target(r=current.rho, phi=current.phi)
-
-
-# -----------------------------------------------
-# 3D
-
-
-@dispatch
-def represent_as(
-    current: CylindricalVector, target: type[Cartesian3DVector], /, **kwargs: Any
-) -> Cartesian3DVector:
-    """CylindricalVector -> Cartesian3DVector."""
-    x = current.rho * xp.cos(current.phi)
-    y = current.rho * xp.sin(current.phi)
-    z = current.z
-    return target(x=x, y=y, z=z)
-
-
-@dispatch
-def represent_as(
-    current: CylindricalVector, target: type[SphericalVector], /, **kwargs: Any
-) -> SphericalVector:
-    """CylindricalVector -> SphericalVector."""
-    r = xp.sqrt(current.rho**2 + current.z**2)
-    theta = xp.acos(current.z / r)
-    phi = current.phi
-    return target(r=r, theta=theta, phi=phi)
